feat(convert.py): add mesh_to_lineset function to convert Open3D mesh to LineSet refactor(test_raycast.py): move mesh_to_lineset function to convert.py and update references refactor(test_raycast.py): rename compute_point_to_mesh_distance to point_to_mesh_distance for consistency
